Fix DIB in SELinux permissive hosts
Remove the DIB element that tries to fix SELinux labels in the build
host after the disk image has been built.

The script fails for many reason when run inside container. We should
not need it because running SELinux in permissive mode and the host
that is being cleaned up is a container.

Change-Id: Iaa3d01b17524a08ee5d2f14e591dcef0dabd3a89
Signed-off-by: Ossi Koivistoinen <ossi.koivistoinen@nokia.com>