Update patch set 1
Patch Set 1:

Same as for ipa-deployer, we're waiting for access to an x86_64 POD for testing.
Jimmy did test an aarch64 ISO with this change and although the deploy did not finish successfully due to some aarch64-specific quirks w.r.t. the partition table, the provisioning did start as expected.

Patch-set: 1
Label: Verified=0