Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)

Hi,
For now the DIB builds on x86_64 explicitly use XFS for the rootfs, so this change (which only applies to UEFI-compatible images, which right now means only aarch64, as x86_64 defaults to block-device-mbr and no UEFI support) tries to sync the BIOS/MBR and UEFI/GPT use-cases partition table format in order to be able to use the same scripts we currently have.

Btw, I haven't added all reviewers to this change as we haven't finished testing the new ISO with UEFI/GPT with XFS, but I'll ping back when we're done.

Patch-set: 1
Label: Verified=0